refactor(cli): split cli.ts into per-group command modules (#48)
`src/cli.ts` had grown to 1388 lines: every command action inline, the
provenance flag trio (`--actor`/`--task-id`/`--reason`) copy-pasted onto
nine commands, the routed-edit flag block duplicated between `route-edit`
and `batch`, and `resolveContent` defined twice.

- `src/cli.ts` is now 196 lines of wiring: root options, the `preAction`
  hook, top-level error handlers, and one `register(program)` call per
  group. Registration order is `--help` order.
- `src/commands/*.ts` — one module per group (read, hash, ast, edit,
  intent, diff, verify, telemetry, provenance, mcp, maintenance, route),
  each exporting `register(program)`. Action bodies moved verbatim.
- `src/commands/shared.ts` — `withProvenance`, `withEditFlags`,
  `withPreview`, `parseRange`, `parseIntFlag`. Helpers wrap a Commander
  chain and append their options last, so `--help` order is unchanged.
- `src/core/resolve-content.ts` — single `resolveContent`, exported from
  the barrel.

No behavior change: `--help` output was byte-compared across the root and
all 40 command paths before and after — identical.

Tests: `tests/cli-structure-48.test.ts` pins the split (cli.ts holds no
`.action(`, every module exports `register`, cli.ts registers all of them,
exactly one `--task-id` declaration, one `resolveContent` and it lives in
core). The two `elapsed_ms` source-scanning tests in `telemetry.test.ts`
and `longtail-51.test.ts` now scan `src/commands/*.ts` plus `cli.ts`, so
they follow the code instead of silently passing on an empty file.
