+++ src/core/index.ts
@@ -135,3 +135,4 @@ export { redactSecrets, redactEvent, isSensitiveFile } from "./redact";
 export { normalizePath, pathsEqual } from "./path-normalize";
 export { decodeText, encodeText, readDecoded } from "./encoding";
 export type { FileEncoding } from "./encoding";
+export { resolveContent } from "./resolve-content";
